third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-garbage-collection-ran-during-error-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
let test_desc = 'Garbage Collection ran during getCharacteristics call that ' +
   'fails. Should not crash';
let service;
promise_test(() => getHealthThermometerService()
  .then(_ => ({service} = _))
  .then(() => service.device.gatt.disconnect())
  .then(runGarbageCollection)
  .then(() => assert_promise_rejects_with_message(
      service.getCharacteristics('measurement_interval'),
      new DOMException(
          'GATT Server is disconnected. Cannot retrieve characteristics. ' +
          '(Re)connect first with `device.gatt.connect`.',
          'NetworkError'))), test_desc);

</script>
